deprecate tooltip link, move pieces into separate files
