+++ frontend/src/component/changeRequest/ChangeRequest/Changes/Change/Change.styles.tsx
@@ -1,0 +1,14 @@
+import { styled, Typography } from '@mui/material';
+
+export const Deleted = styled(Typography)(({ theme }) => ({
+    color: theme.palette.error.main,
+    '::before': { content: '"- "' },
+}));
+
+export const Added = styled(Typography)(({ theme }) => ({
+    '::before': { content: '"+ "' },
+}));
+
+export const Added = styled(Typography)(({ theme }) => ({
+    '::before': { content: '"+ "' },
+}));
